Peeling Inaccuracy
Peeling has a +25 to hit, right now.
This swaps that to a -10. Simple as.

We were discussing behind the scenes a way to make weapons more viable, for those that don't have access to peeling. Given right now, a sabre or rapier outperforms most weapons because of peeling being a wild thing.

This may well be a way to do it. Is it good? I 'unno. Peeling is weird and awful though, but we're stuck with it. May as well make the best of a bad thing.

Doubly so without gutting existing weaponry. Ironic as it is to say, given recent PRs.