Peel Annihilation
Removes all current instances of peel, outright, as used on weaponry.
This doesn't remove the mechanic code wise, or the intents themselves, as they still exist if we wish to use it in the future. Should someone wish to justify the return. As messy as it is, for it be be slinking about.

For not-so-great compensation, it provides striking with a blade 10% more integrity damage, and mordhau bashing 30% more.