Garrote Buff - Absolver Rites - Slurbow Rarity - - - Does as it says on the title. A number of changes facing Inquisition content.
Firstly, garrote.
These required EIGHT resists to snap. You would never actually knock someone out with this, as it has 200 integrity, and takes 25 damage per resist. On escape, it immediately broke the wire.
I've replaced this with removal of the integrity damage on resist, with escapes dealing a flat 100 integrity damage. Two escapes will snap it. I hope that it adds some actual use to an otherwise typically ignored item, given it's only used currently for stopping chases and muting mages, as opposed to the actual functionality intended for it. Especially since wrestling 'balance' as is make these even more painful to use.

Skipping ahead, slurbows.
I've removed them from the merchant vendor, fullstop. They had an identical price range to crossbows. For some reason. Wow. Wonder why. Additionally, knaves get it at an identical price setup to the longbow, making it a direct upgrade of the crossbow. Because it is.

Now, to actual content. Absolver rites.
This takes the rite circle stuff, the existing non-functional rite circle for Psydonites and actually turns it into something with content.
Three rites exist. 'Conversion', 'Admonishment' and 'Freedom'. 'Admonishment' and 'Freedom' provide a five minute cooldown. 'Conversion' can be used at any time, so long as that cooldown isn't in effect.