Increases Direbears Con to 12
After rigorous testing; this was proven to not be as OP as feared; and when/if the direbear changes back; 12 con will actually likely insta-crit or knock the druid into dead; meaning it enforces they remain in the shape as a tank.

Damage is perfect as it is; no changes needed.