Mercs Use Wretch Slot Setup
Mercenaries now make use of the wretch slot setup, so they scale with pop. As opposed to 8 roundstart, in an eleven player round.

They start at 3 slots. For every ten players above a count of twenty, it goes up by one, back to the original 8 slots. Is this harsh? Yeah. But it should prevent eight man merc deathstacks who outbid the keep's own roundstart.

I would be more harsh with this, but this is going to have enough pushback as is from the grenzel stacks and folks who play nothing but Ruma / Shaman / whatever to get their meta setup roundstart.